+++ api/routes/analysis.py
@@ -16,10 +16,10 @@ from api.utils.db import (
     get_projectless_db_connection,
 )
 from api.utils.export import ExportType
+from db.python.db_filters import GenericFilter
 from db.python.layers.analysis import AnalysisLayer
 from db.python.tables.analysis import AnalysisFilter
 from db.python.tables.project import ProjectPermissionsTable
-from db.python.utils import GenericFilter
 from models.enums import AnalysisStatus
 from models.models.analysis import (
     Analysis,
