+++ db/python/db_filters/__init__.py
@@ -1,0 +1,8 @@
+# This file is used to import all the filters that are used in the application
+
+from db.python.db_filters.generic import (
+    GenericFilter,
+    GenericFilterModel,
+    GenericMetaFilter,
+    get_hashable_value,
+)
