+++ db/python/layers/cohort.py
@@ -1,4 +1,5 @@
 from db.python.connect import Connection
+from db.python.db_filters import GenericFilter
 from db.python.layers.base import BaseLayer
 from db.python.layers.sequencing_group import SequencingGroupLayer
 from db.python.tables.cohort import CohortFilter, CohortTable, CohortTemplateFilter
@@ -8,7 +9,7 @@ from db.python.tables.sequencing_group import (
     SequencingGroupFilter,
     SequencingGroupTable,
 )
-from db.python.utils import GenericFilter, get_logger
+from db.python.utils import get_logger
 from models.models.cohort import (
     CohortCriteriaInternal,
     CohortInternal,
