+++ db/python/tables/analysis_runner.py
@@ -1,8 +1,9 @@
 import dataclasses
 import datetime
 
+from db.python.db_filters import GenericFilter, GenericFilterModel
 from db.python.tables.base import DbBase
-from db.python.utils import GenericFilter, GenericFilterModel, to_db_json
+from db.python.utils import to_db_json
 from models.models.analysis_runner import AnalysisRunnerInternal
 from models.models.project import ProjectId
