+++ db/python/tables/cohort.py
@@ -2,9 +2,10 @@
 import dataclasses
 import datetime
 
+from db.python.db_filters import GenericFilter, GenericFilterModel
 from db.python.tables.base import DbBase
 from db.python.tables.project import ProjectId
-from db.python.utils import GenericFilter, GenericFilterModel, NotFoundError, to_db_json
+from db.python.utils import NotFoundError, to_db_json
 from models.models.cohort import (
     CohortCriteriaInternal,
     CohortInternal,
