+++ test/test_participant.py
@@ -1,8 +1,8 @@
 from test.testbase import DbIsolatedTest, run_as_sync
 
+from db.python.db_filters import GenericFilter
 from db.python.layers.participant import ParticipantLayer
 from db.python.tables.participant import ParticipantFilter
-from db.python.utils import GenericFilter
 from models.models.assay import AssayUpsertInternal
 from models.models.participant import ParticipantUpsertInternal
 from models.models.sample import SampleUpsertInternal
