+++ test/test_project_groups.py
@@ -4,10 +4,9 @@ from test.testbase import DbIsolatedTest, run_as_sync
 from db.python.tables.project import (
     GROUP_NAME_MEMBERS_ADMIN,
     GROUP_NAME_PROJECT_CREATORS,
-    Forbidden,
     ProjectPermissionsTable,
 )
-from db.python.utils import NotFoundError
+from db.python.utils import Forbidden, NotFoundError
 
 
 class TestGroupAccess(DbIsolatedTest):
